+++ src/main/java/kishida/cnn/activation/RetifierdLinear.java
@@ -9,12 +9,12 @@ package kishida.cnn.activation;
 public class RetifierdLinear extends ActivationFunction {
 
     @Override
-    public double apply(double value) {
+    public float apply(float value) {
         return value >= 0 ? value : 0;
     }
 
     @Override
-    public double diff(double value) {
+    public float diff(float value) {
         return value >= 0 ? 1 : 0;
     }
