+++ src/main/java/kishida/cnn/layers/MultiNormalizeLayer.java
@@ -13,34 +13,34 @@ import kishida.cnn.activation.LinearFunction;
  * @author naoki
  */
 public class MultiNormalizeLayer extends ImageNeuralLayer{
-    public MultiNormalizeLayer(String name, int size, double threshold, ImageNeuralLayer preLayer, boolean useGpu) {
+    public MultiNormalizeLayer(String name, int size, float threshold, ImageNeuralLayer preLayer, boolean useGpu) {
         this(name, preLayer.outputChannels, preLayer.outputWidth, preLayer.outputHeight, preLayer, size, threshold, useGpu);
     }
 
 
-    public MultiNormalizeLayer(String name, int inputChannels, int inputWidth, int inputHeight, ImageNeuralLayer preLayer, int size, double threshold, boolean useGpu) {
+    public MultiNormalizeLayer(String name, int inputChannels, int inputWidth, int inputHeight, ImageNeuralLayer preLayer, int size, float threshold, boolean useGpu) {
         super(name, new LinearFunction(), inputChannels, inputWidth, inputHeight, inputChannels, inputWidth, inputHeight);
         this.preLayer = preLayer;
         this.size = size;
         this.threshold = threshold;
         this.useGpu = useGpu;
-        averages = new double[inputWidth * inputHeight];
-        rates = new double[inputWidth * inputHeight];
-        result = new double[inputChannels * inputHeight * inputWidth];
+        averages = new float[inputWidth * inputHeight];
+        rates = new float[inputWidth * inputHeight];
+        result = new float[inputChannels * inputHeight * inputWidth];
     }
 
-    double[] averages;
-    double[] rates;
+    float[] averages;
+    float[] rates;
     int size;
-    double threshold;
+    float threshold;
     boolean useGpu;
 
     @Override
-    public double[] forward(double[] in) {
+    public float[] forward(float[] in) {
 
         IntStream.range(0, inputWidth).parallel().forEach(x -> {
             for(int y = 0; y < inputHeight; ++y){
-                double total = 0;
+                float total = 0;
                 int count = 0;
                 for(int i = 0; i < size; ++i){
                     int xx = x + i - size / 2;
@@ -58,8 +58,8 @@ public class MultiNormalizeLayer extends ImageNeuralLayer{
                         }
                     }
                 }
-                double average = total / count;
-                double variance = 0;
+                float average = total / count;
+                float variance = 0;
                 for(int i = 0; i < size; ++i){
                     int xx = x + i - size / 2;
                     if(xx < 0 || xx >= inputWidth){
@@ -71,12 +71,12 @@ public class MultiNormalizeLayer extends ImageNeuralLayer{
                             continue;
                         }
                         for(int ch = 0; ch < inputChannels; ++ch){
-                            double data = in[ch * inputHeight * inputWidth + xx * inputHeight + yy];
+                            float data = in[ch * inputHeight * inputWidth + xx * inputHeight + yy];
                             variance += (data - average) * (data - average);
                         }
                     }
                 }
-                double std = Math.max(threshold, Math.sqrt(variance / count));
+                float std = Math.max(threshold, (float)Math.sqrt(variance / count));
                 averages[x * inputHeight + y] = average;
                 rates[x * inputHeight + y] = std;
                 for(int ch = 0; ch < inputChannels; ++ch){
@@ -90,7 +90,7 @@ public class MultiNormalizeLayer extends ImageNeuralLayer{
     }
 
     @Override
-    public double[] backward(double[] in, double[] delta) {
+    public float[] backward(float[] in, float[] delta) {
         return delta;
     }
